% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics of neural network}
\usage{
diagnostics(predictions, y.test)
}
\arguments{
\item{predictions}{Vector of labels typically obtained as output of \code{predict} function.}

\item{y.test}{Vector of labels typically from the testing dataset.}
}
\value{
A list containing
\itemize{
 \item \code{Accuracy}
 \item \code{F1.scores}
}
}
\description{
Evaluate performance parameters of a neural network.
}
\details{
Calculates accuracy and calculates F1 score by evaluating precision and recall. \code{predications} and \code{y.test} must be vectors of same length.
}
\examples{
y <- data.matrix(iris)[1:100,5]
x <- data.matrix(iris)[101:150,1:4]
t <- nnlearn(x,y,layers = 2,hidden.units = 3,lambda = 0.85)
p <- nnpredict(t,x,y)
y.test <- data.matrix(iris)[101:150,5]
D <- diagnostics(p,y.test)
}
