% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnpredict.R
\name{nnpredict}
\alias{nnpredict}
\title{Predict labels}
\usage{
nnpredict(weights, x.new, y.train)
}
\arguments{
\item{weights}{List of weight matrices. Typically obtained as the output of \code{nnlearn} function.}

\item{x.new}{Matrix of feature data typically from the testing dataset.}

\item{y.train}{Vector of labels typically from the training dataset.}
}
\value{
Vector containing predicted labels corresponding to rows of \code{x.new}.
}
\description{
Predict labels for new feature data.
}
\examples{
y <- data.matrix(iris)[1:100,5]
x <- data.matrix(iris)[101:150,1:4]
t <- nnlearn(x,y,layers = 2,hidden.units = 3,lambda = 0.85)
p <- nnpredict(t,x,y)
}
