% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{Matrix containing feature data.}
}
\value{
Normalized form of the input matrix with the same dimensions.
}
\description{
Normalize skewed data before training neural network.
}
\details{
Normalizes data to have zero mean by subtracting column-wise mean from data and the scales data by dividing with standard deviation.
\deqn{x_{norm} = \frac{x-\mu_{x}}{\sigma_{x}}}{xnorm = (x-\mu)/\sigma}
}
\examples{
x <- data.matrix(iris)[,1:4]
x.norm <- normalize(x)
}
