% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_init.R
\name{rand_init}
\alias{rand_init}
\title{Random Initialization}
\usage{
rand_init(L_in, L_out, epsilon.init = (sqrt(6)/sqrt(L_in + L_out)))
}
\arguments{
\item{L_in}{Positive integer equal to the number of neurons in the previous/incoming hidden layer.}

\item{L_out}{Positive integer equal to the number of neurons in the next/outgoing hidden layer.}

\item{epsilon.init}{Real number parameter for symmetry breaking.}
}
\value{
A matrix of size \code{L_out by (1+L_in)}
}
\description{
Randomly initialize weight matrices while implementing symmetry breaking.
}
\details{
If the value of epsilon is not specified, it takes on the default value of \code{sqrt(6)/sqrt(L_in + L_out)} since this is considered a suitable value for \eqn{\epsilon}.
}
\examples{
t <- rand_init(3,3,0.15)
th <- rand_init(4,5)
}
