% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(m)
}
\arguments{
\item{m}{Matrix or vector input.}
}
\value{
Matrix with dimensions same as the input matrix/vector.
}
\description{
Evaluate element-wise sigmoid function of a matrix.
}
\details{
Sigmoid function is the default activation function contained in each neuron of the neural network.
\deqn{sigmoid(m) = \frac{1}{1+e^{-m}}}{sigmoid(m) = 1/(1+exp(-m))}
The \code{sigmoid} function will be used very prominently in the forward pass of the neural networks algorithm.
}
\examples{
m <- c(1:30)
s <- sigmoid(m)
x <- matrix(c(1:45),15,3)
y <- sigmoid(x)
}
