% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidgradient.R
\name{sigmoidgradient}
\alias{sigmoidgradient}
\title{Gradient of sigmoid function}
\usage{
sigmoidgradient(z)
}
\arguments{
\item{z}{Matrix/vector input}
}
\value{
Matrix with dimensions same as the input matrix.
}
\description{
Evaluate the gradient of the sigmoid function.
}
\details{
\code{sigmoidgradient} implements \code{sigmoid} function. The \code{sigmoidgradient} function is a significant part of the backpropagation part of the neural network algorithm.
}
\examples{
m <- matrix(c(1:60),4,15)
sg <- sigmoidgradient(m)
}
