% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttsplit.R
\name{ttsplit}
\alias{ttsplit}
\title{Test-Train split}
\usage{
ttsplit(x, y, fraction = 0.66)
}
\arguments{
\item{x}{Matrix containing feature data.}

\item{y}{Vector of labels corresponding to feature data contained in \code{x}. \code{y} must be the same length as \code{x}.}

\item{fraction}{Real number between 0 and 1.}
}
\value{
A list containing
\describe{
 \item{\code{x.train}}{Feature data to be used for training.}
 \item{\code{y.train}}{Labels corresponding to \code{x.train}.}
 \item{\code{x.test}}{Feature data to be used for testing.}
 \item{\code{y.test}}{Labels corresponding to \code{x.test}.} 
}
}
\description{
Split the data into training and testing datasets.
}
\details{
Function samples random indices and creates sub datasets for training and testing from the original complete dataset.
If the value of \code{fraction} in unspecified, it takes on the default value 0.66.
}
\examples{
data <- data.matrix(iris)
x <- data[,1:4]
y <- data[,5]
nn.data <- ttsplit(x,y)
nn.data <- ttsplit(x,y,fraction = 0.70)
}
